Changed f_jk
The issue was using the same residual for both the sums. Due to the same residual the run for j=0 was ending up fine as we ran the inside for loop only once. But if j was bigger than 0 then every time the inner for loop ran, the residual would record the value UP to that particular l instead of just at that l. Then it would add on itself, counting the zeroth value l times, first value l-1 times and so on. this was the over-counting error we were observing for the case of P_{11} at F_{0}^{7}q_1. I've corrected that by introducing an inner residual so now the code works alright.